requirements.txt
cycler==0.11.0
fonttools==4.28.5
kiwisolver==1.3.2
numpy==1.22.1
packaging==21.3
Pillow==9.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
six==1.16.0
